fixed a test

+++ src/test/java/RewardValueTests.java
@@ -13,7 +13,7 @@ public class RewardValueTests {
 
     @Test
     void create_with_miles_value() {
-        int milesValue = 10000;
+        Double milesValue = 10000.0;
         var rewardValue = new RewardValue(milesValue);
         assertEquals(milesValue, rewardValue.getMilesValue());
     }
